Refactored ECS into Core and ECS definition; Added main thread support
